tools/macos-companion/requirements.txt
Pillow==10.3.0
opencv-python==4.9.0.80
mediapipe==0.10.21
flask==3.0.3
numpy==1.26.4 